azure-functions/transcribe-audio/__init__.py
import azure.functions as func
import azure.cognitiveservices.speech as speechsdk
import json
import os
import logging

def main(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request.')

    try:
        # Get the request body
        req_body = req.get_json()
        audio_data = req_body.get('audio')
        
        if not audio_data:
            return func.HttpResponse(
                "Please pass audio data in the request body",
                status_code=400
            )

        # Get Azure Speech credentials from environment variables
        speech_key = os.environ["AZURE_SPEECH_KEY"]
        speech_region = os.environ["AZURE_SPEECH_REGION"]

        # Configure speech service
        speech_config = speechsdk.SpeechConfig(
            subscription=speech_key, 
            region=speech_region
        )
        speech_config.speech_recognition_language = "en-US"

        # Process the audio and get transcription
        # Note: This is a simplified version. We'll need to modify this
        # to handle the actual audio data format and processing
        result = {
            "transcription": "Sample transcription",
            "speakers": []
        }

        return func.HttpResponse(
            json.dumps(result),
            mimetype="application/json",
            status_code=200
        )

    except Exception as e:
        logging.error(f"Error processing request: {str(e)}")
        return func.HttpResponse(
            f"Error processing request: {str(e)}",
            status_code=500
        ) 