azure-functions/requirements.txt
azure-cognitiveservices-speech==1.31.0
azure-communication-callautomation==1.0.0
python-dotenv==1.0.0
websockets==12.0
azure-functions==1.17.0
azure-storage-blob==12.17.0
numpy==1.24.3 